+++ crates/rust-client/src/rpc/web_tonic_client/mod.rs
@@ -16,13 +16,13 @@ use miden_objects::{
     Digest,
 };
 use miden_tx::utils::Serializable;
+use tonic::Streaming;
 use tonic_web_wasm_client::Client;
 
 use super::{
     domain::{
         account::{AccountDetails, AccountProof, AccountProofs, AccountUpdateSummary},
         note::{NetworkNote, NoteSyncInfo},
-        sync::StateSyncInfo,
     },
     generated::{
         requests::{
@@ -30,6 +30,7 @@ use super::{
             GetAccountProofsRequest, GetBlockHeaderByNumberRequest, GetNotesByIdRequest,
             SubmitProvenTransactionRequest, SyncNoteRequest, SyncStateRequest,
         },
+        responses::SyncStateResponse,
         rpc::api_client::ApiClient,
     },
     NodeRpcClient, NodeRpcClientEndpoint, RpcError,
@@ -182,7 +183,7 @@ impl NodeRpcClient for WebTonicRpcClient {
         account_ids: &[AccountId],
         note_tags: &[NoteTag],
         nullifiers_tags: &[u16],
-    ) -> Result<StateSyncInfo, RpcError> {
+    ) -> Result<Streaming<SyncStateResponse>, RpcError> {
         let mut query_client = self.build_api_client();
 
         let account_ids = account_ids.iter().map(|acc| (*acc).into()).collect();
@@ -199,7 +200,7 @@ impl NodeRpcClient for WebTonicRpcClient {
         let response = query_client.sync_state(request).await.map_err(|err| {
             RpcError::RequestError(NodeRpcClientEndpoint::SyncState.to_string(), err.to_string())
         })?;
-        response.into_inner().try_into()
+        Ok(response.into_inner())
     }
 
     async fn sync_notes(