+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * Add check for empty pay to id notes (#714).
+* Read Sync State response from stream (#734).
 
 ## 0.7.0 (2025-01-28)